refactor code

+++ app/src/main/java/sondo65/com/androidtest/utils/Constants.java
@@ -3,6 +3,6 @@ package sondo65.com.androidtest.utils;
 public class Constants {
 
     //If you want to change DEFINED_ROW value.Clear data or reinstall app to see update!
-    public static final int DEFINED_ROW = 272128;
+    public static final double DEFINED_ROW = 272128;
     public static final int LOAD_MORE_ITEM = 30;
 }
